getting and setting cover is working tests pass
